DEV: Prepare plugin for deprecation.
We'll merge this plugin into core soon and need to make a couple of changes to avoid conflict. Changes are:

- Wrap the sitemap model in a module so core can define its own.
- Don't update sitemaps or add the path to the robots.txt file if the `publish_sitemap` setting is present, which is what's core is going to use.
- Use append instead of prepend for mounting the engine, so the same routes defined by core can have a higher priority, and we never hit this plugin's controller.
- Core will use the same table if present, so we add `if_not_exists: true` in the migration just in case conflicts can originate based on the call order.
