+++ app/jobs/scheduled/update_sitemaps.rb
@@ -5,7 +5,9 @@ module ::Jobs
     every 1.hour
 
     def execute(args)
-      Sitemap.update! if SiteSetting.sitemap_enabled
+      return if SiteSetting.respond_to?(:publish_sitemaps)
+
+      DiscourseSitemap::Sitemap.update! if SiteSetting.sitemap_enabled
     end
   end
 end
