+++ app/views/connectors/robots_txt_index/sitemap.html.erb
@@ -1,5 +1,6 @@
-<%- if SiteSetting.sitemap_enabled? %>
-<%- unless SiteSetting.login_required? %>
+
+<%- if !SiteSetting.respond_to?(:publish_sitemaps) %>
+<%- if SiteSetting.sitemap_enabled? && !SiteSetting.login_required? %>
 Sitemap: <%= request.protocol %><%= request.host_with_port %>/sitemap.xml
 <% end %>
 <% end %>
