+++ db/migrate/20200909134134_create_sitemaps.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-class CreateSitemaps < ActiveRecord::Migration[6.0]
+class CreateSitemaps < ActiveRecord::Migration[6.1]
   def change
-    create_table :sitemaps do |t|
+    create_table :sitemaps, if_not_exists: true do |t|
       t.string :name, null: false
       t.datetime :last_posted_at, null: false
       t.boolean :enabled, null: false, default: true
     end
 
-    add_index :sitemaps, :name, unique: true
+    add_index :sitemaps, :name, unique: true, if_not_exists: true
   end
 end
