+++ spec/requests/sitemap_controller_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 RSpec.describe DiscourseSitemap::SitemapController do
   describe '#default' do
     before do
-      Sitemap.create!(name: '1', last_posted_at: Time.now)
+      DiscourseSitemap::Sitemap.create!(name: '1', last_posted_at: Time.now)
     end
 
     it 'does not fail then page is a string starting with a number' do
